Added UserGroupController and moved association-permission endpoint to it
