+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/routing/web/WebUserRouteConfig.kt
@@ -9,12 +9,6 @@ object WebUserRouteConfig : RouteConfig
 {
     private val usersManage = setOf("*/users.manage")
     override val endpoints = listOf(
-            WebEndpoint("/users/:email/actions/associate-permission/",
-                    UserController::class, "associatePermission",
-                        contentType = ContentTypes.json,
-                        method = HttpMethod.post)
-                    .json()
-                    .secure(usersManage),
             WebEndpoint("/users/report-readers/:report/",
                 UserController::class, "getReportReaders",
                     contentType = ContentTypes.json)
