+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/controllers/web/UserGroupController.kt
@@ -1,0 +1,43 @@
+package org.vaccineimpact.orderlyweb.controllers.web
+
+import java.net.URLDecoder
+import org.vaccineimpact.orderlyweb.ActionContext
+import org.vaccineimpact.orderlyweb.controllers.Controller
+import org.vaccineimpact.orderlyweb.db.AuthorizationRepository
+import org.vaccineimpact.orderlyweb.db.OrderlyAuthorizationRepository
+import org.vaccineimpact.orderlyweb.models.Scope
+import org.vaccineimpact.orderlyweb.models.permissions.AssociatePermission
+import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
+
+class UserGroupController(context: ActionContext,
+                     val authRepo : AuthorizationRepository) : Controller(context)
+{
+    constructor(context: ActionContext) : this(context, OrderlyAuthorizationRepository())
+
+    fun associatePermission(): String
+    {
+        val userGroupId = userGroupId()
+
+        val postData = context.postData()
+        val associatePermission = AssociatePermission(
+                postData["action"]!!,
+                postData["name"]!!,
+                postData["scope_prefix"],
+                postData["scope_id"]
+        )
+
+        val permission = ReifiedPermission(associatePermission.name, Scope.parse(associatePermission))
+
+        when (associatePermission.action)
+        {
+            "add" -> authRepo.ensureUserGroupHasPermission(userGroupId, permission)
+            "remove" -> authRepo.ensureUserGroupDoesNotHavePermission(userGroupId, permission)
+            else -> throw IllegalArgumentException("Unknown action type")
+        }
+
+        return okayResponse()
+    }
+
+
+    private fun userGroupId(): String = URLDecoder.decode(context.params(":user-group-id"), "UTF-8")
+}
