+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/unit_tests/controllers/web/UserGroupControllerTests.kt
@@ -1,0 +1,101 @@
+package org.vaccineimpact.orderlyweb.tests.unit_tests.controllers.web
+
+import com.nhaarman.mockito_kotlin.*
+import org.assertj.core.api.Assertions
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.Test
+import org.mockito.ArgumentCaptor
+import org.vaccineimpact.orderlyweb.ActionContext
+import org.vaccineimpact.orderlyweb.controllers.web.UserGroupController
+import org.vaccineimpact.orderlyweb.db.AuthorizationRepository
+import org.vaccineimpact.orderlyweb.db.UserRepository
+import org.vaccineimpact.orderlyweb.models.User
+import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
+import org.vaccineimpact.orderlyweb.test_helpers.TeamcityTests
+import java.lang.IllegalArgumentException
+import java.time.Instant
+
+class UserGroupControllerTests : TeamcityTests()
+{
+    @Test
+    fun `adds permission to user group`()
+    {
+        val actionContext = mock<ActionContext> {
+            on { this.params(":email") } doReturn "user1%40example.com"
+            on { this.postData() } doReturn mapOf(
+                    "action" to "add",
+                    "name" to "test.permission",
+                    "scope_prefix" to "report",
+                    "scope_id" to "report1"
+            )
+        }
+
+        val authRepo = mock<AuthorizationRepository>()
+        val sut = UserGroupController(actionContext, authRepo)
+        val result = sut.associatePermission()
+
+        assertThat(result).isEqualTo("OK")
+
+        val permissionCaptor: ArgumentCaptor<ReifiedPermission> = ArgumentCaptor.forClass(ReifiedPermission::class.java)
+        verify(authRepo).ensureUserGroupHasPermission(eq("user1@example.com"), capture(permissionCaptor))
+
+        val permission = permissionCaptor.value
+        Assertions.assertThat(permission.name).isEqualTo("test.permission")
+        Assertions.assertThat(permission.scope.value).isEqualTo("report:report1")
+    }
+
+    @Test
+    fun `removes permission from user group`()
+    {
+        val actionContext = mock<ActionContext> {
+            on { this.params(":email") } doReturn "user1%40example.com"
+            on { this.postData() } doReturn mapOf(
+                    "action" to "remove",
+                    "name" to "test.permission",
+                    "scope_prefix" to "report",
+                    "scope_id" to "report1"
+            )
+        }
+
+        val authRepo = mock<AuthorizationRepository>()
+        val sut = UserGroupController(actionContext, authRepo)
+        val result = sut.associatePermission()
+
+        assertThat(result).isEqualTo("OK")
+
+        val permissionCaptor: ArgumentCaptor<ReifiedPermission> = ArgumentCaptor.forClass(ReifiedPermission::class.java)
+        verify(authRepo).ensureUserGroupDoesNotHavePermission(eq("user1@example.com"), capture(permissionCaptor))
+
+        val permission = permissionCaptor.value
+        Assertions.assertThat(permission.name).isEqualTo("test.permission")
+        Assertions.assertThat(permission.scope.value).isEqualTo("report:report1")
+    }
+
+    @Test
+    fun `throws exception if associate permission with unknown action`()
+    {
+        val actionContext = mock<ActionContext> {
+            on { this.params(":email") } doReturn "user1%40example.com"
+            on { this.postData() } doReturn mapOf(
+                    "action" to "addx",
+                    "name" to "test.permission",
+                    "scope_prefix" to "report",
+                    "scope_id" to "report1"
+            )
+        }
+
+        val userRepo = mock<UserRepository> {
+            on { this.getUser("user1@example.com") } doReturn User("user.1",
+                    "User One",
+                    "user1@example.com",
+                    "test",
+                    Instant.now())
+        }
+
+        val authRepo = mock<AuthorizationRepository>()
+        val sut = UserGroupController(actionContext, authRepo)
+        Assertions.assertThatThrownBy { sut.associatePermission() }.isInstanceOf(IllegalArgumentException::class.java)
+
+    }
+
+}
